extract action_type mod.
