+++ types/src/account.rs
@@ -1,6 +1,7 @@
 //! Contains types and constants associated with user accounts.
 
 mod account_hash;
+mod action_type;
 mod weight;
 
 use crate::BLAKE2B_DIGEST_LENGTH;
@@ -25,6 +26,7 @@ pub use self::{
         AccountHash, AccountHashBytes, TryFromSliceForAccountHashError,
         ACCOUNT_HASH_FORMATTED_STRING_PREFIX, ACCOUNT_HASH_LENGTH,
     },
+    action_type::ActionType,
     weight::{Weight, WEIGHT_SERIALIZED_LENGTH},
 };
 
@@ -64,34 +66,6 @@ impl Display for FromStrError {
                 write!(f, "failed to decode address portion from hex: {}", error)
             }
             FromStrError::Hash(error) => write!(f, "address portion is wrong length: {}", error),
-        }
-    }
-}
-
-/// The various types of action which can be performed in the context of a given account.
-#[repr(u32)]
-pub enum ActionType {
-    /// Represents performing a deploy.
-    Deployment = 0,
-    /// Represents changing the associated keys (i.e. map of [`AccountHash`]s to [`Weight`]s) or
-    /// action thresholds (i.e. the total [`Weight`]s of signing [`AccountHash`]s required to
-    /// perform various actions).
-    KeyManagement = 1,
-}
-
-// This conversion is not intended to be used by third party crates.
-#[doc(hidden)]
-impl TryFrom<u32> for ActionType {
-    type Error = TryFromIntError;
-
-    fn try_from(value: u32) -> Result<Self, Self::Error> {
-        // This doesn't use `num_derive` traits such as FromPrimitive and ToPrimitive
-        // that helps to automatically create `from_u32` and `to_u32`. This approach
-        // gives better control over generated code.
-        match value {
-            d if d == ActionType::Deployment as u32 => Ok(ActionType::Deployment),
-            d if d == ActionType::KeyManagement as u32 => Ok(ActionType::KeyManagement),
-            _ => Err(TryFromIntError(())),
         }
     }
 }
